+++ src/hashes/CHash1.h
@@ -50,7 +50,7 @@ public:
 	CHash1(unsigned int, unsigned int);
 	~CHash1();
     
-    unsigned short int hash(byte* buffer, unsigned int offset, unsigned int length);
+    unsigned int hash(byte* buffer, unsigned int offset, unsigned int length);
 };
 
 #endif
