+++ src/hashes/DefaultHash.cpp
@@ -40,7 +40,7 @@ DefaultHash::DefaultHash(unsigned int hashSize, unsigned int maxWordLength) : Ha
 DefaultHash::~DefaultHash() {
 }
 
-unsigned short int DefaultHash::hash(byte* buffer, unsigned int offset, unsigned int length) {
+unsigned int DefaultHash::hash(byte* buffer, unsigned int offset, unsigned int length) {
     unsigned int hash = 0;
 	unsigned int limit = length <= maxWordLength ? length : maxWordLength;
     for(unsigned int i = 0; i < limit; i ++)
