+++ src/hashes/DefaultHash.h
@@ -40,7 +40,7 @@ public:
 	DefaultHash(unsigned int, unsigned int);
 	~DefaultHash();
 
-    unsigned short int hash(byte* buffer, unsigned int offset, unsigned int length);
+    unsigned unsigned int hash(byte* buffer, unsigned int offset, unsigned int length);
 };
 
 #endif
