+++ src/hashes/SdbmHash.cpp
@@ -38,7 +38,7 @@ SdbmHash::SdbmHash(unsigned int hashSize, unsigned int maxWordLength) : HashFunc
 SdbmHash::~SdbmHash() {
 }
 
-unsigned short int SdbmHash::hash(byte* buffer, unsigned int offset, unsigned int length) {
+unsigned int SdbmHash::hash(byte* buffer, unsigned int offset, unsigned int length) {
     unsigned int hash = 0;
 	//unsigned int limit = length <= maxWordLength ? length : maxWordLength;
     for(unsigned int i = 0; i < length; i ++)
